Removed grid-table and refactor on GridComponent to improve implementation.
