--- src/lib/src/grid/tb-column.html
@@ -1,4 +1,0 @@
-<ng-container cdkColumnDef="{{columnDef}}">
-    <cdk-header-cell *cdkHeaderCellDef> {{header}} </cdk-header-cell>
-    <cdk-cell *cdkCellDef="let row"> {{row[columnDeff]}} </cdk-cell>
-</ng-container>
